Make CFeeRate work with uint64_t sizes

+++ src/policy/feerate.cpp
@@ -9,7 +9,7 @@
 
 #include <cmath>
 
-CFeeRate::CFeeRate(const CAmount& nFeePaid, uint32_t num_bytes)
+CFeeRate::CFeeRate(const CAmount& nFeePaid, uint64_t num_bytes)
 {
     const int64_t nSize{num_bytes};
 
@@ -20,7 +20,7 @@ CFeeRate::CFeeRate(const CAmount& nFeePaid, uint32_t num_bytes)
     }
 }
 
-CAmount CFeeRate::GetFee(uint32_t num_bytes) const
+CAmount CFeeRate::GetFee(uint64_t num_bytes) const
 {
     const int64_t nSize{num_bytes};
